+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Support to nullable page properties (#149)
+  - Properties: `Date`, `Email`, `Number`, `PhoneNumber`, `Select`, `Url`.
+  - New method `isEmpty()` on those properties.
+
 ## [1.1.0]
 
 ## Added